@@ -1,13 +1,14 @@
 {
   "name": "@orh/vue-kityminder",
-  "description": "Vue 2 kityminder",
+  "description": "Vue kityminder",
   "keywords": [
     "vue",
     "vue 2",
+    "vue 3",
     "kityminder",
     "vue-kityminder"
   ],
-  "version": "1.3.0",
+  "version": "2.0.0",
   "main": "./src/index.js",
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -15,7 +16,7 @@
     "lint": "vue-cli-service lint"
   },
   "peerDependencies": {
-    "vue": "^2.6.11"
+    "vue": "^3.0.0"
   },
   "dependencies": {
     "kity": "^2.0.4",
@@ -25,14 +26,14 @@
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "babel-eslint": "^10.1.0",
-    "core-js": "^3.6.5",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
+    "eslint-plugin-vue": "^7.0.0",
     "sass": "^1.26.5",
     "sass-loader": "^8.0.2",
-    "vue": "^2.6.11",
-    "vue-template-compiler": "^2.6.11"
+    "core-js": "^3.6.5",
+    "vue": "^3.0.0"
   },
   "eslintConfig": {
     "root": true,
@@ -40,7 +41,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
