@@ -1,11 +1,5 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import VueKityminder from './index'
 
-Vue.use(VueKityminder);
-
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).use(VueKityminder).mount('#app')
