@@ -1,7 +1,7 @@
 import VueKityminder from './components/VueKityminder';
 
-const install = Vue => {
-  Vue.component(VueKityminder.name, VueKityminder);
-};
-
-export default {install};
+export default {
+  install: app => {
+    app.component(VueKityminder.name, VueKityminder);
+  }
+}
